refactor: relocate effect classes into src/Effects domain
