--- src/Actions/Effects/RedirectEffect.php
@@ -1,15 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Lattice\Lattice\Actions\Effects;
-
-use Lattice\Lattice\Actions\Enums\EffectType;
-use Lattice\Lattice\Attributes;
-
-#[Attributes\AsEffect(EffectType::Redirect)]
-final readonly class RedirectEffect extends AbstractEffect
-{
-    public function __construct(
-        public string $url,
-    ) {}
-}
