--- src/Actions/Effects/ReloadPageEffect.php
@@ -1,10 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Lattice\Lattice\Actions\Effects;
-
-use Lattice\Lattice\Actions\Enums\EffectType;
-use Lattice\Lattice\Attributes;
-
-#[Attributes\AsEffect(EffectType::ReloadPage)]
-final readonly class ReloadPageEffect extends AbstractEffect {}
