+++ src/Effects/Builtin/ReloadComponentEffect.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+namespace Lattice\Lattice\Effects\Builtin;
+
+use Lattice\Lattice\Effects\Attributes\AsEffect;
+use Lattice\Lattice\Effects\Effect;
+use Lattice\Lattice\Effects\Enums\EffectType;
+
+#[AsEffect(EffectType::ReloadComponent)]
+final readonly class ReloadComponentEffect extends Effect
+{
+    public function __construct(
+        public string $component,
+    ) {}
+}
